+++ wayland-client/examples/simple_window.rs
@@ -1,4 +1,4 @@
-use std::{fs::File, os::unix::prelude::AsRawFd};
+use std::{fs::File, os::unix::prelude::AsFd};
 
 use wayland_client::{
     delegate_noop,
@@ -73,8 +73,7 @@ impl Dispatch<wl_registry::WlRegistry, ()> for State {
 
                     let mut file = tempfile::tempfile().unwrap();
                     draw(&mut file, (init_w, init_h));
-                    let pool =
-                        shm.create_pool(file.as_raw_fd(), (init_w * init_h * 4) as i32, qh, ());
+                    let pool = shm.create_pool(file.as_fd(), (init_w * init_h * 4) as i32, qh, ());
                     let buffer = pool.create_buffer(
                         0,
                         init_w as i32,